[Bronze V] Title: 이상한 기호, Time: 100 ms, Memory: 14192 KB, Score: 100 point -BaekjoonHub

+++ 백준/Bronze/15964. 이상한 기호/이상한 기호.java
@@ -16,7 +16,7 @@ public class Main
         A = Integer.parseInt(st.nextToken());
         B = Integer.parseInt(st.nextToken());
 
-        long answer = (A + B) * (A - B);
+        long answer = (long) (A + B) * (A - B);
 
         System.out.println(answer);
     }
